requirements.txt
streamlit
PyPDF2
langchain
langchain-classic
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
python-dotenv

